+++ lib/qc/modules/qc_tester/audit/components/audit_question_summary_component.dart
@@ -1,14 +1,17 @@
+import 'dart:async';
+
 import 'package:builder_component/builder_component.dart';
 import 'package:calculator_ui/calculator_ui.dart';
 import 'package:core_widgets/core_widgets.dart';
 import 'package:csh_annotation/annotation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/audit/resources/new_audit_response.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/home/screens/qc_tester_home_screen.dart';
 import 'package:flutter_trc/src/app_builder/app_builder_groups/groups.dart';
 import 'package:flutter_trc/src/app_builder/app_headers/general_app_header/models/none_config_model.dart';
 import 'package:provider/provider.dart';
 
 import '../l10n.dart';
-import '../models/audit_question_response.dart';
 import '../models/audit_question_summary_comp_param.dart';
 import '../providers/audit_submission_provider.dart';
 import '../screens/audit_barcode_scanner_screen.dart';
@@ -55,7 +58,7 @@ class AuditQuestionSummaryComponent extends StatelessComponent<NoneConfigModel> 
   }
 }
 
-class _AuditSummary extends StatelessWidget {
+class _AuditSummary extends StatefulWidget {
   final AuditQuestionResponse? dataModel;
 
   final String scannedBarcode;
@@ -67,11 +70,25 @@ class _AuditSummary extends StatelessWidget {
   }) : super(key: key);
 
   @override
+  State<_AuditSummary> createState() => _AuditSummaryState();
+}
+
+class _AuditSummaryState extends State<_AuditSummary> {
+  @override
+  void initState() {
+    if (!Validator.isListNullOrEmpty(widget.dataModel?.manualAuditQuestionList)) {
+      scheduleMicrotask(() {
+        _openManualAuditQuestionModal(widget.dataModel!.manualAuditQuestionList!);
+      });
+    }
+    super.initState();
+  }
+
+  @override
   Widget build(BuildContext context) {
     var l10n = L10n(context);
     var theme = Theme.of(context);
-    var provider = AuditQuestionSubmitProvider.of(context);
-    return (dataModel != null && !Validator.isListNullOrEmpty(dataModel!.questionList))
+    return (widget.dataModel != null && !Validator.isListNullOrEmpty(widget.dataModel!.auditQuestionList))
         ? Column(
             children: [
               Expanded(
@@ -79,11 +96,11 @@ class _AuditSummary extends StatelessWidget {
                   itemBuilder: (context, index) {
                     return SubmittedQuestionWidget(
                       questionNumber: index,
-                      question: dataModel!.questionList![index].question ?? "",
-                      answeredQuestion: dataModel!.questionList![index].selectedOption ?? "",
+                      question: widget.dataModel!.auditQuestionList![index].question ?? "",
+                      answeredQuestion: widget.dataModel!.auditQuestionList![index].selectedOption ?? "",
                     );
                   },
-                  itemCount: dataModel!.questionList!.length,
+                  itemCount: widget.dataModel!.auditQuestionList!.length,
                 ),
               ),
               Padding(
@@ -111,8 +128,8 @@ class _AuditSummary extends StatelessWidget {
 
   _createPostDataMap() {
     Map<String, dynamic> postDataMap = {};
-    if (!Validator.isListNullOrEmpty(dataModel?.questionList)) {
-      for (var element in dataModel!.questionList!) {
+    if (!Validator.isListNullOrEmpty(widget.dataModel?.auditQuestionList)) {
+      for (var element in widget.dataModel!.auditQuestionList!) {
         if (element.selectedOption != null && !Validator.isListNullOrEmpty(element.options!.values.toList())) {
           String key = element.options!.keys
               .firstWhere((item) => element.options![item] == element.selectedOption, orElse: () => "");
@@ -133,7 +150,7 @@ class _AuditSummary extends StatelessWidget {
   _submitQuestionsResults(L10n l10n, BuildContext context, Map<String, dynamic> dataMap) {
     var provider = AuditQuestionSubmitProvider.of(context, listen: false);
     CshLoading().showLoading(context);
-    provider.submitAuditQuestion(scannedBarcode, dataMap).then((value) {
+    provider.submitAuditQuestion(widget.scannedBarcode, dataMap).then((value) {
       CshLoading().hideLoading(context);
       CshSnackBar.success(
         context: context,
@@ -149,7 +166,7 @@ class _AuditSummary extends StatelessWidget {
 
   _getDeviceStatus(AuditQuestionSubmitProvider provider, BuildContext context) {
     CshLoading().showLoading(context);
-    provider.getDeviceStatus(scannedBarcode).then((value) {
+    provider.getDeviceStatus(widget.scannedBarcode).then((value) {
       CshLoading().hideLoading(context);
       _showDeviceStatusDialog(context, value);
     }, onError: (e) {
@@ -161,7 +178,8 @@ class _AuditSummary extends StatelessWidget {
   _showDeviceStatusDialog(BuildContext context, String deviceStatus) {
     showAlertDialog(context, title: "Channel", desc: deviceStatus, onPosBtnPressed: (_) {
       Navigator.pop(context); // dismiss dialog
-      Navigator.of(context).pushNamedAndRemoveUntil(AuditBarcodeScannerScreen.route, (route) => false);
+      Navigator.of(context).pushNamedAndRemoveUntil(
+          AuditBarcodeScannerScreen.route, (route) => route.settings.name == QcTesterHomeScreen.route);
     });
   }
 
@@ -194,5 +212,60 @@ class _AuditSummary extends StatelessWidget {
             ],
           ),
         ));
+  }
+
+  void _openManualAuditQuestionModal(List<ManualAuditQuestionItem> questionList) {
+    showCshBottomSheet(
+      isDismissible: false,
+      context: context,
+      child: PopScope(
+        canPop: false,
+        child: StatefulBuilder(builder: (_, setState) {
+          return Container(
+            height: MediaQuery.of(context).size.height * 0.8,
+            padding: const EdgeInsets.all(Dimens.space_16),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                CshTextNew.h4("Verify these questions"),
+                const SizedBox(height: Dimens.space_24),
+                Expanded(
+                  child: ListView.separated(
+                    itemCount: questionList.length,
+                    itemBuilder: (_, index) {
+                      var item = questionList[index];
+                      return CshCheckbox(
+                        title: CshTextNew.subTitle1(item.question ?? ""),
+                        isSelected: item.isSelected ?? false,
+                        onChanged: (value) {
+                          setState(() {
+                            item.isSelected = value;
+                          });
+                        },
+                      );
+                    },
+                    separatorBuilder: (BuildContext context, int index) {
+                      return const SizedBox(height: Dimens.space_16);
+                    },
+                  ),
+                ),
+                const SizedBox(height: Dimens.space_16),
+                Center(
+                  child: CshBigButton(
+                    text: "Proceed",
+                    onPressed: () {
+                      var provider = AuditQuestionSubmitProvider.of(context, listen: false);
+                      Navigator.pop(context); // Dismiss dialog
+                      provider.onManualQuestionAnswered(questionList);
+                    },
+                  ),
+                )
+              ],
+            ),
+          );
+        }),
+      ),
+    );
   }
 }