+++ lib/qc/modules/qc_tester/lob_devices/resources/device_detail_response.g.dart
@@ -6,30 +6,19 @@ part of 'device_detail_response.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-DeviceDetailResponse _$DeviceDetailResponseFromJson(
-        Map<String, dynamic> json) =>
-    DeviceDetailResponse(
-      json['dt'] == null
-          ? null
-          : DeviceDetailResponseData.fromJson(
-              json['dt'] as Map<String, dynamic>),
-      json['__ca'] == null
-          ? null
-          : CashifyAlert.fromJson(json['__ca'] as Map<String, dynamic>),
+DeviceDetailResponse _$DeviceDetailResponseFromJson(Map<String, dynamic> json) => DeviceDetailResponse(
+      json['dt'] == null ? null : DeviceDetailResponseData.fromJson(json['dt'] as Map<String, dynamic>),
+      json['__ca'] == null ? null : CashifyAlert.fromJson(json['__ca'] as Map<String, dynamic>),
       json['turl'] as String?,
     );
 
-Map<String, dynamic> _$DeviceDetailResponseToJson(
-        DeviceDetailResponse instance) =>
-    <String, dynamic>{
+Map<String, dynamic> _$DeviceDetailResponseToJson(DeviceDetailResponse instance) => <String, dynamic>{
       '__ca': instance.cashifyAlert,
       'turl': instance.trackUrl,
       'dt': instance.deviceDetails,
     };
 
-DeviceDetailResponseData _$DeviceDetailResponseDataFromJson(
-        Map<String, dynamic> json) =>
-    DeviceDetailResponseData(
+DeviceDetailResponseData _$DeviceDetailResponseDataFromJson(Map<String, dynamic> json) => DeviceDetailResponseData(
       json['qc'] as String?,
       json['imei'] as String?,
       json['sn'] as String?,
@@ -38,15 +27,17 @@ DeviceDetailResponseData _$DeviceDetailResponseDataFromJson(
         (k, e) => MapEntry(int.parse(k), e as String),
       ),
       json['imei2'] as String?,
+      (json['rm'] as Map<String, dynamic>?)?.map(
+                (k, e) => MapEntry(k, e as int),
+      ),
     );
 
-Map<String, dynamic> _$DeviceDetailResponseDataToJson(
-        DeviceDetailResponseData instance) =>
-    <String, dynamic>{
+Map<String, dynamic> _$DeviceDetailResponseDataToJson(DeviceDetailResponseData instance) => <String, dynamic>{
       'qc': instance.qrCode,
       'imei': instance.imei1,
       'imei2': instance.imei2,
       'sn': instance.serialNo,
       'cat_id': instance.selectedCategoryId,
       'cat': instance.categories?.map((k, e) => MapEntry(k.toString(), e)),
+      'rm': instance.reasons?.map((k, e) => MapEntry(k, e.toString())),
     };