+++ lib/qc/modules/qc_tester/lob_devices/resources/reasons.dart
@@ -1,0 +1,6 @@
+class Reasons {
+  String? name;
+  int? isImageRequired;
+
+  Reasons(this.name, this.isImageRequired);
+}